requirements.txt
asgiref==3.4.1
certifi==2021.10.8
Django==3.2.5
djangorestframework==3.12.4
pip==21.2.4
psycopg2==2.8.6
Pygments==2.10.0
pytz==2021.3
setuptools==58.0.4
sqlparse==0.4.1
typing-extensions==3.10.0.2
wheel==0.37.0
wincertstore==0.2
